requirements.txt
fire>=0.3.1
fontTools>=4.38.0
freetype_py>=2.3.0
setuptools>=62.6.0
